Gets the links of the movies now
Altering the code in a way to open the webpage with selenium, scroll to the bottom of the webpage and then take the page source to the html parser. After that with the help of regex it finds the href attributes of the found a elements of the code. Then turns them into links and saves them in a list.